+++ compages/_structure.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 from collections.abc import Callable, Iterable, Mapping
-from typing import Any, NewType, TypeVar, get_origin, overload
+from typing import Any, NewType, TypeVar, overload
 
-from ._common import GeneratorStack
+from ._common import GeneratorStack, get_lookup_order
 from .path import PathElem
 
 
@@ -33,17 +33,25 @@ def collect_messages(
     return result
 
 
+class SequentialStructureHandler(ABC):
+    @abstractmethod
+    def applies(self, structure_into: Any, val: Any) -> bool: ...
+
+    @abstractmethod
+    def __call__(self, structurer: "Structurer", structure_into: Any, val: Any) -> Any: ...
+
+
 _T = TypeVar("_T")
 
 
 class Structurer:
     def __init__(
         self,
-        handlers: Mapping[Any, Callable[["Structurer", type, Any], Any]] = {},
-        predicate_handlers: Iterable["PredicateStructureHandler"] = [],
+        lookup_handlers: Mapping[Any, Callable[["Structurer", type, Any], Any]] = {},
+        sequential_handlers: Iterable[SequentialStructureHandler] = [],
     ):
-        self._handlers = handlers
-        self._predicate_handlers = predicate_handlers
+        self._lookup_handlers = lookup_handlers
+        self._sequential_handlers = sequential_handlers
 
     @overload
     def structure_into(self, structure_into: NewType, val: Any) -> Any: ...
@@ -53,31 +61,19 @@ class Structurer:
 
     def structure_into(self, structure_into: Any, val: Any) -> Any:
         stack = GeneratorStack((self, structure_into), val)
+        lookup_order = get_lookup_order(structure_into)
 
-        # First check if there is an exact match registered
-        handler = self._handlers.get(structure_into, None)
-        if stack.push(handler):
-            return stack.result()
-
-        # If it's a newtype, try to fall back to a handler for the wrapped type
-        if isinstance(structure_into, NewType):
-            handler = self._handlers.get(structure_into.__supertype__, None)
+        for tp in lookup_order:
+            handler = self._lookup_handlers.get(tp, None)
             if stack.push(handler):
                 return stack.result()
 
-        # If it's a generic, see if there is a handler for the generic origin
-        origin = get_origin(structure_into)
-        if origin is not None:
-            handler = self._handlers.get(origin, None)
-            if stack.push(handler):
-                return stack.result()
-
-        # Check all predicate handlers in order and see if there is one that applies
+        # Check all sequential handlers in order and see if there is one that applies
         # TODO (#10): should `applies()` raise an exception which we could collect
         # and attach to the error below, to provide more context on why no handlers were found?
-        for predicate_handler in self._predicate_handlers:
-            if predicate_handler.applies(structure_into, val):
-                if stack.push(predicate_handler):
+        for sequential_handler in self._sequential_handlers:
+            if sequential_handler.applies(structure_into, val):
+                if stack.push(sequential_handler):
                     return stack.result()
                 break
 
@@ -87,11 +83,3 @@ class Structurer:
         raise StructuringError(
             f"Could not find a non-generator handler to structure into {structure_into}"
         )
-
-
-class PredicateStructureHandler(ABC):
-    @abstractmethod
-    def applies(self, structure_into: Any, val: Any) -> bool: ...
-
-    @abstractmethod
-    def __call__(self, structurer: Structurer, structure_into: Any, val: Any) -> Any: ...