+++ tests/test_structure_handlers.py
@@ -35,7 +35,7 @@ def assert_exception_matches(exc, reference_exc):
 
 
 def test_structure_into_none():
-    structurer = Structurer(handlers={type(None): structure_into_none})
+    structurer = Structurer(lookup_handlers={type(None): structure_into_none})
     assert structurer.structure_into(type(None), None) is None
 
     with pytest.raises(StructuringError) as exc:
@@ -45,7 +45,7 @@ def test_structure_into_none():
 
 
 def test_structure_into_float():
-    structurer = Structurer(handlers={float: structure_into_float})
+    structurer = Structurer(lookup_handlers={float: structure_into_float})
     assert structurer.structure_into(float, 1.5) == 1.5
 
     # Specifically allow integers, but check that they are converted to floats.
@@ -60,7 +60,7 @@ def test_structure_into_float():
 
 
 def test_structure_into_bool():
-    structurer = Structurer(handlers={bool: structure_into_bool})
+    structurer = Structurer(lookup_handlers={bool: structure_into_bool})
     assert structurer.structure_into(bool, True) is True
     assert structurer.structure_into(bool, False) is False
 
@@ -71,7 +71,7 @@ def test_structure_into_bool():
 
 
 def test_structure_into_str():
-    structurer = Structurer(handlers={str: structure_into_str})
+    structurer = Structurer(lookup_handlers={str: structure_into_str})
     assert structurer.structure_into(str, "abc") == "abc"
 
     with pytest.raises(StructuringError) as exc:
@@ -81,7 +81,7 @@ def test_structure_into_str():
 
 
 def test_structure_into_bytes():
-    structurer = Structurer(handlers={bytes: structure_into_bytes})
+    structurer = Structurer(lookup_handlers={bytes: structure_into_bytes})
     assert structurer.structure_into(bytes, b"abc") == b"abc"
 
     with pytest.raises(StructuringError) as exc:
@@ -91,7 +91,7 @@ def test_structure_into_bytes():
 
 
 def test_structure_into_int():
-    structurer = Structurer(handlers={int: structure_into_int})
+    structurer = Structurer(lookup_handlers={int: structure_into_int})
     assert structurer.structure_into(int, 1) == 1
 
     with pytest.raises(StructuringError) as exc:
@@ -109,7 +109,11 @@ def test_structure_into_int():
 
 def test_structure_into_union():
     structurer = Structurer(
-        handlers={UnionType: structure_into_union, int: structure_into_int, str: structure_into_str}
+        lookup_handlers={
+            UnionType: structure_into_union,
+            int: structure_into_int,
+            str: structure_into_str,
+        }
     )
     assert structurer.structure_into(int | str, "a") == "a"
     assert structurer.structure_into(int | str, 1) == 1
@@ -128,7 +132,11 @@ def test_structure_into_union():
 
 def test_structure_into_tuple():
     structurer = Structurer(
-        handlers={tuple: structure_into_tuple, int: structure_into_int, str: structure_into_str}
+        lookup_handlers={
+            tuple: structure_into_tuple,
+            int: structure_into_int,
+            str: structure_into_str,
+        }
     )
 
     assert structurer.structure_into(tuple[()], []) == ()
@@ -161,7 +169,7 @@ def test_structure_into_tuple():
 
 
 def test_structure_into_list():
-    structurer = Structurer(handlers={list: structure_into_list, int: structure_into_int})
+    structurer = Structurer(lookup_handlers={list: structure_into_list, int: structure_into_int})
 
     assert structurer.structure_into(list[int], [1, 2, 3]) == [1, 2, 3]
     assert structurer.structure_into(list[int], (1, 2, 3)) == [1, 2, 3]
@@ -182,7 +190,11 @@ def test_structure_into_list():
 
 def test_structure_into_dict():
     structurer = Structurer(
-        handlers={dict: structure_into_dict, int: structure_into_int, str: structure_into_str}
+        lookup_handlers={
+            dict: structure_into_dict,
+            int: structure_into_int,
+            str: structure_into_str,
+        }
     )
 
     assert structurer.structure_into(dict[int, str], {1: "a", 2: "b"}) == {1: "a", 2: "b"}
@@ -213,8 +225,8 @@ def test_structure_into_dict():
 
 def test_structure_dataclass_from_list():
     structurer = Structurer(
-        handlers={int: structure_into_int, str: structure_into_str},
-        predicate_handlers=[StructureListIntoDataclass()],
+        lookup_handlers={int: structure_into_int, str: structure_into_str},
+        sequential_handlers=[StructureListIntoDataclass()],
     )
 
     @dataclass
@@ -250,8 +262,8 @@ def test_structure_dataclass_from_list():
 
 def test_structure_dataclass_from_dict():
     structurer = Structurer(
-        handlers={int: structure_into_int, str: structure_into_str},
-        predicate_handlers=[
+        lookup_handlers={int: structure_into_int, str: structure_into_str},
+        sequential_handlers=[
             StructureDictIntoDataclass(name_converter=lambda name, _metadata: name + "_")
         ],
     )
@@ -287,8 +299,8 @@ def test_structure_dataclass_from_dict():
 
     # Need a structurer without a name converter for this one
     structurer = Structurer(
-        handlers={int: structure_into_int, str: structure_into_str},
-        predicate_handlers=[StructureDictIntoDataclass()],
+        lookup_handlers={int: structure_into_int, str: structure_into_str},
+        sequential_handlers=[StructureDictIntoDataclass()],
     )
     with pytest.raises(StructuringError) as exc:
         structurer.structure_into(Container, {"x": 1, "z": "b"})