+++ docs/changelog.rst
@@ -1,15 +1,29 @@
 Changelog
 =========
 
-0.2.1 (2024-03-05)
+0.3.0 (unreleased)
 ------------------
+
+Changed
+^^^^^^^
+
+- Renamed "handlers" and "predicate handlers" to "lookup handlers" and "sequential handlers". (PR_15_)
+
 
 Added
 ^^^^^
 
 - Skip fields equal to the defaults when unstructuring dataclasses. (PR_13_)
 - Generator-based deferring to lower level structuring and unstructuring. (PR_13_)
+- Support for ``NewType`` chains. (PR_15_)
 
+
+.. _PR_13: https://github.com/fjarri/compages/pull/13
+.. _PR_15: https://github.com/fjarri/compages/pull/15
+
+
+0.2.1 (2024-03-05)
+------------------
 
 Fixed
 ^^^^^
@@ -18,7 +32,6 @@ Fixed
 
 
 .. _PR_1: https://github.com/fjarri/compages/pull/1
-.. _PR_13: https://github.com/fjarri/compages/pull/13
 
 
 0.2.0 (2024-03-05)